*: Revamp codename
This keeps the codename Oklo, and sets the stage for
this variable to not be buried in this shell script.
Instead providing it through the `manifest/version.txt` as
`FLATCAR_VERSION_CODENAME`.
If that value is present in the version.txt, then it will now clobber
this burried `OS_CODENAME`.

Signed-off-by: Vincent Batts <vbatts@kinvolk.io>