sys-kernel/coreos-sources: Drop duplicates from configs
Most of the duplicates happened to be within amd64 config. But some of
the duplicates were across the files (defined in both common and
amd64). Almost all of them were exact duplicates, so those could be
just dropped, favoring the config lines in common config to remain.

The CONFIG_CONNECTOR was different in common config (module) and in
amd64 config (builtin), so the config line was moved to arm64
config. Now amd64 has CONFIG_CONNECTOR as builtin and arm64 as a
module.