app-emulation/containerd: add containerd-runc-shim-v*
Add new binaries containerd-runc-shim-v[12] to the torcx tarballs for
docker and containerd. The binaries are necessary for kubelet to
communicate via custom CRI endpoints.

The addition will cause usage of the /usr partition to grow by ~5M.

```
$ ls -l /run/torcx/unpack/docker/bin
-rwxr-xr-x. 1 root root  6742592 Sep 30 13:22 containerd-shim
-rwxr-xr-x. 1 root root  9095176 Sep 30 13:22 containerd-shim-runc-v1
-rwxr-xr-x. 1 root root  9111752 Sep 30 13:22 containerd-shim-runc-v2

$ ls -l /usr/share/torcx/store/docker\:19.03.torcx.tgz
-rw-r--r--. 1 root root 89809888 Sep 30 14:16 /usr/share/torcx/store/docker:19.03.torcx.tgz

$ df /usr
Filesystem      1K-blocks   Used Available Use% Mounted on
/dev/mapper/usr   1007760 916024     39720  96% /usr
```

Note, we do not touch other torcx profiles like docker 1.12 or 17.03,
to keep the image size as small as possible.