build_library: Do not try to upload update zip archive
The generate_update function only generates two files, a .bin file and
a .gz file, so these two should be uploaded. A .zip file is generated
by the zip_update_tools function, which already takes care of the
uploading part (apparently without signing, which is done
separately?). Confusingly the .zip file does not contain the update
payload, but rather some tools for applying the update.