profiles: Disable split-usr explicitly for generic images
We don't want to have separate /bin, /lib, /usr/bin and /usr/lib
directories. The former two are meant to be symlinks to the latter
two. The `split-usr` USE flag gets enabled with the profiles update in
portage-stable, so before doing the update, clear the flag in the
overlay.

This is not done for SDK images, since they seem to have split /usr on
purpose.