coreos-base/hard-host-depends: bring back dependency on dev-libs/nspr
Now that `dev-libs/nss` is removed from the depencencies list of
hard-host-depends, SDK does not include `dev-libs/nspr` any more.
As a result, `dev-lang/spidermonkey` fails to build, because it requires
`dev-libs/nspr` in the SDK. It is not sufficient to have nspr under
`/build/amd64-usr`.

Add `dev-libs/nspr` back to the dependencies of `hard-host-depends`,
to make it included in the SDK.