coreos-base/gmerge: Stop installing gmerge script
It is not used anywhere during the build process, thus drop
it. Dropping it makes it easier to port this ebuild to python3, since
there will be only one script to port to python3. The
`emerge-gitclone` script will need porting anyway, because it imports
portage code, which will become python3 after the update.

Most likely the package should be then renamed to
`coreos-base/emerge-gitclone`, but this can be done later.