app-emulation/containerd: delete containerd-stress
Each Flatcar production image includes a binary `containerd-stress`,
as a part of torcx tarballs.

However it does not seem to be used anywhere.
It looks like a stress testing tool for containerd, so I don't see a
good reason to keep it.
The binary was there since the beginning, via commit
[fdd926949a10](https://github.com/kinvolk/coreos-overlay/commit/fdd926949a10c3011746a57a04fc36f451fb8009),
but there is no comment or messages why it was needed.
We can simply remove `containerd-stress`.