net-fs/samba: new USE flags spotlight, regedit, glusterfs, ntvfs
Introduce a USE flag spotlight, to be able to disable the spotlight
backend by default, as it is not needed by Linux.

Introduce a USE flag rededit, to be able to disable the rededit
tool if needed.

Introduce a USE flag glusterfs, to be able to disable the glusterfs
by default.

Introduce a USE flag ntvfs, to be able to disable the ntvfs-fileserver
by default.

Since the docbook-xsl-stylesheets and libxslt are needed only
at build time, we should move those deps to BDEPEND.