amd64 SDK: add cros_host to BOOTSTRAP_USE
This change adds the USE flag cros_host to the
SDK's make.default, as part of a larger fix for the SDK bootstrap build.
The SDK bootstrap build was broken in stage 1 since package upgrades
were allowed to leak into that phase.

We now limit stage 1 to only "known good" package ebuilds, which caused
downstream breakage from missing flags in the stage 2 SDK bootstrapping.

This change fixes that breakage.

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>