app-crypt/mit-krb5: add warn-crossbuild.patch
Set PYTHON_COMPAT to python 3.6 and 3.7 to be suitable for the current
code base.

Add a custom patch to replace error with warning when running autoconf
for cross builds, because libkrb5 is not able to detect
cross-compilation.

Based on 64e33c9f826d8fd951fd58ba1ed70debaf65be8d .