coreos-base/update_engine: add glib-utils to (B)DEPEND
glib-utils are used during the build so they need to be part of host
dependencies for update_engine. This only really pops up during a repeat
bootstrap, when update_engine is being built from source but glib has
been installed from a binary. BDEPEND would be the correct variable but
that requires EAPI=7, so additionally added it to DEPEND for now.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>