coreos-base/oem-azure: refactor grub.cfg to apply console override only on PC
We always want the 'flatcar.autologin' parameter, but the ttyS0 setting is x86
PC specific. Move the generic part to linux_append variable and hide the
generic part behind a check for grub_platform. For EFI platforms the default
grub.cfg has the correct arch specific console configuration.

The console specification for grub itself is needed in either case

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>