net-misc/openssh: sync with Gentoo for 8.7_p1-r1
Update net-misc/openssh to 8.7_p1-r1, mainly to address CVE-2020-15778.

Goal of the package update is to add the support of a new option `-s`
of scp, i.e. "sftp mode of scp". Openssh 8.7 started to support the
flag, but it is disabled by default. So at the moment users need to
explicitly run `scp -s` to test the feature.

Gentoo ref: 11d6f23704e7ab84191e28e034816bfdb151d406