script/bootstrap: update openssl before stage3
Right now our bootstrap flow is different then gentoo's - we don't
update the seed when building stage1 and use a different ebuilds
snapshot for stage1 compared to stage2 and stage3. This is causing us
trouble now, because we introduced openssl-3, but seed/stage1 still
contains openssl-1.1. During `emerge -e @system` in stage3, some
packages that depend on openssl may build against the stage1 version,
which results in an error during depcleaning (they would need to be
rebuilt instead).  Stage3 is not extensible, so instead, explicitly
update openssl in stage2. This workaround can be removed as soon as we
release a seed with openssl-3.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>