ci-automation + setup_board: publish and use binpkgs
It is a back-port of the 3 commits in
https://github.com/flatcar-linux/scripts/pull/209
to flatcar-3033. The backport was required because of a conflict
in set_version.

This change adds a job for publishing binary packages to the build cache
server to the ci automation.

Also, setup_board is updated to use the buildcache package cache if a
nightly build version is detected.

Lastly, ci-automation/README.md is updated to reflect the changes. THe
readme now also elaborates on the binary package cache server (build
infrastructure).

Co-authored-by: Krzesimir Nowak <knowak@microsoft.com>