build_library/check_root: ignore dangling '*egg-info' symlink
Since the update of dev-python/certifi, running the command
`./image_to_vm.sh --format gce --board=amd64-usr` fails due to a
dangling symlink. This symlink is located in
/usr/lib64/python3.9/site-packages and is not supposed to be installed
in the first place because of this INSTALL_MASK entry in
coreos-overlay/profiles/coreos/targets/generic/oem-aci/make.defaults:

  INSTALL_MASK="${INSTALL_MASK}
    /usr/*/python3*
  "

There is an open upstream bug that INSTALL_MASK doesn't work correctly on
symlinks (https://bugs.gentoo.org/678462).
The best we can do at this time is to ignore the dangling symlink.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>